Вносит правки согласно замечаниям 9.1
